implemented basic authentication to update or delete users

+++ pastebin/users/views.py
@@ -4,9 +4,11 @@ import transaction
 from starlette.endpoints import HTTPEndpoint
 from starlette.requests import Request
 from starlette.responses import JSONResponse, PlainTextResponse
+from starlette.authentication import requires
 
 from pastebin.mixins import SAModelMixin
-from .models import User
+from pastebin.settings import DEFAULT_USER_GROUP
+from .models import User, Group
 
 
 class Users(HTTPEndpoint):
@@ -33,6 +35,8 @@ class Users(HTTPEndpoint):
         password = payload.pop('password')
         user = User(**payload)
         user.set_password(password)
+        group = db.query(Group).filter_by(name=DEFAULT_USER_GROUP).one()
+        user.groups.append(group)
         db.add(user)
         db.flush()
         data = {
@@ -60,11 +64,16 @@ class UserInfo(SAModelMixin, HTTPEndpoint):
             'created_at': str(user.created_at)
         })
 
+    @requires(['authenticated', 'users:write'])
     async def put(self, request: Request) -> JSONResponse:
         user = self.get_model_by_id(request, User, request.path_params['id'])
+        self.check_ownership(request, user)
         payload = await request.json()
-        for item in ['created_at', 'updated_at', 'id', 'password']:
+        for item in ['created_at', 'updated_at', 'id']:
             payload.pop(item, None)
+        password = payload.pop('password', None)
+        if password is not None:
+            user.set_password(password)
 
         for key, value in payload.items():
             setattr(user, key, value)
@@ -80,8 +89,10 @@ class UserInfo(SAModelMixin, HTTPEndpoint):
             'created_at': str(user.created_at)
         })
 
+    @requires(['authenticated', 'users:write'])
     def delete(self, request: Request) -> PlainTextResponse:
         user = self.get_model_by_id(request, User, request.path_params['id'])
+        self.check_ownership(request, user)
         request.state.db.delete(user)
         transaction.commit()
         return PlainTextResponse('', status_code=204)
